Move some utils to shared folder
