+++ src/extension/providers/dart_diagnostic_provider.ts
@@ -1,8 +1,8 @@
 import { Diagnostic, DiagnosticCollection, DiagnosticRelatedInformation, DiagnosticSeverity, DiagnosticTag, Location, Uri } from "vscode";
 import * as as from "../../shared/analysis_server_types";
+import { toRangeOnLine } from "../../shared/vscode/utils";
 import { Analyzer } from "../analysis/analyzer";
 import { config } from "../config";
-import { toRangeOnLine } from "../utils";
 
 // TODO: This is not a provider?
 export class DartDiagnosticProvider {
