+++ src/extension/providers/dart_document_symbol_provider.ts
@@ -2,9 +2,9 @@ import { CancellationToken, DocumentSymbol, DocumentSymbolProvider, TextDocument
 import { Outline } from "../../shared/analysis_server_types";
 import { Logger } from "../../shared/interfaces";
 import { waitFor } from "../../shared/utils/promises";
+import { toRange } from "../../shared/vscode/utils";
 import { getSymbolKindForElementKind } from "../analysis/analyzer_das";
 import { openFileTracker } from "../analysis/open_file_tracker";
-import { toRange } from "../utils";
 
 export class DartDocumentSymbolProvider implements DocumentSymbolProvider {
 	constructor(private readonly logger: Logger) { }
