+++ src/extension/providers/dart_workspace_symbol_provider.ts
@@ -3,10 +3,9 @@ import { CancellationToken, Location, SymbolInformation, Uri, workspace, Workspa
 import * as as from "../../shared/analysis_server_types";
 import { Logger } from "../../shared/interfaces";
 import { escapeRegExp } from "../../shared/utils";
-import { fsPath } from "../../shared/vscode/utils";
+import { fsPath, toRange } from "../../shared/vscode/utils";
 import { Analyzer } from "../analysis/analyzer";
 import { getSymbolKindForElementKind } from "../analysis/analyzer_das";
-import { toRange } from "../utils";
 
 export class DartWorkspaceSymbolProvider implements WorkspaceSymbolProvider {
 	private badChars: RegExp = new RegExp("[^0-9a-z\-]", "gi");
