+++ src/shared/vscode/utils.ts
@@ -1,5 +1,6 @@
-import { ExtensionKind, extensions, Uri, workspace, WorkspaceFolder } from "vscode";
+import { ExtensionKind, extensions, Position, Range, TextDocument, Uri, workspace, WorkspaceFolder } from "vscode";
 import { dartCodeExtensionIdentifier } from "../constants";
+import { Location } from "../interfaces";
 import { forceWindowsDriveLetterToUppercase } from "../utils";
 
 const dartExtension = extensions.getExtension(dartCodeExtensionIdentifier);
@@ -27,4 +28,20 @@ export function isDartWorkspaceFolder(folder?: WorkspaceFolder): boolean {
 	// For now, nothing calls this that will do bad things if the folder isn't a Dart
 	// project so we can review amend this in future if required.
 	return true;
+}
+
+export function toRange(document: TextDocument, offset: number, length: number): Range {
+	return new Range(document.positionAt(offset), document.positionAt(offset + length));
+}
+
+export function toPosition(location: Location): Position {
+	return new Position(location.startLine - 1, location.startColumn - 1);
+}
+
+// Translates an offset/length to a Range.
+// NOTE: Does not wrap lines because it does not have access to a TextDocument to know
+// where the line ends.
+export function toRangeOnLine(location: Location): Range {
+	const startPos = toPosition(location);
+	return new Range(startPos, startPos.translate(0, location.length));
 }
